4.25

+++ source/isaaclab_tasks/isaaclab_tasks/direct/ergonomic_hrta/eg_hrta_task_manager.py
@@ -73,7 +73,7 @@ class TaskManager(object):
             self.fatigue_data[task] = copy.deepcopy(self.obs) 
             self.fatigue_data[task]['phy_fatigue'] = torch.tensor([self.characters.fatigue_list[charac_idx].phy_fatigue], dtype=torch.float32)
             self.fatigue_data[task]['psy_fatigue'] = torch.tensor([self.characters.fatigue_list[charac_idx].psy_fatigue], dtype=torch.float32)
-            self.fatigue_data[task]['charac_idx'] = torch.tensor(charac_idx, dtype=torch.int32) 
+            self.fatigue_data[task]['charac_idx'] = torch.tensor(charac_idx, dtype=torch.int64) 
             self.fatigue_data[task]['task_str'] = task
             self.fatigue_data[task]['action'] = torch.tensor(self.task_dic_inverse[task]+1, dtype=torch.int32)
             # self.fatigue_data[task]['prediction_mask'] = torch.zeros((len(self.task_dic), 2), dtype=torch.float32)
